+++ libs/langchain-mongodb/tests/utils.py
@@ -72,7 +72,7 @@ class PatchedMongoDBAtlasVectorSearch(MongoDBAtlasVectorSearch):
         if self._is_autoembedding:
             while monotonic() - start <= TIMEOUT:
                 for idx in list(self.collection.list_search_indexes()):
-                    if idx["name"] == "langchain-test-index-from-texts-autoEmbed":
+                    if idx["name"] == self._index_name:
                         if idx["numDocs"] == n_docs:
                             return ids_inserted
                 sleep(INTERVAL)